Fix episode with question marks, and improve !ep detector
Question marks need to be encoded in the URL because they
would otherwise be interpreted as a URL query.

Also, the `!ep` detector was a little broken: The `elsif`
had no effect because anything that begins with `!ep` must
also begin with `!episode`.

Bonus: Ignore extra text after the episode number, so you can say

    !ep 714 This draft episode was oddly prescient.

and the bot will ignore your commentary.